Rename diagnostics to analytics and update imports
Replaces 'diagnostics' terminology with 'analytics' across experiment parameter files and updates related variable names and toggles. Renames wave_diagnostics.py to L0_analytics.py and updates imports and usage in L0_launcher.py. Removes unused diagnostics imports and code from radial_wave archive scripts for consistency.
